+++ src/UserBugs/mocks/bugs.ts
@@ -1,0 +1,23 @@
+export const bugs = {
+  morningBug: 4,
+  nightBug: 3,
+  goldenBug: 100
+};
+
+export const bugsArray: {
+  type: 'morningBug' | 'nightBug' | 'goldenBug';
+  color: string;
+}[] = [
+  {
+    type: 'morningBug',
+    color: 'text-light-point'
+  },
+  {
+    type: 'nightBug',
+    color: 'text-dark-point'
+  },
+  {
+    type: 'goldenBug',
+    color: 'text-warning'
+  }
+];
