+++ src/UserProfile/mocks/userData.ts
@@ -1,0 +1,18 @@
+export const userData = {
+  nickname: '자려고 사는 사람',
+  intro: '너무나 졸린사람',
+  level: 3,
+  exp: 6,
+  birds: {
+    MORNING_SKIN:
+      'https://i.pinimg.com/564x/9d/1d/81/9d1d81196806e7a2829f94f67ccc3248.jpg',
+    NIGHT_SKIN:
+      'https://i.pinimg.com/564x/f4/9d/21/f49d217cae2f110e8acee4ed1f120483.jpg'
+  },
+  badges: {
+    오목눈이_탄생: true,
+    어른_오목눈이: false,
+    어른_부엉이: false,
+    부엉이_탄생: true
+  }
+};
